Add ability to set max URLs to return
So as to allow for the download-then-delete process to be chunked, add the ability
to specify the maximum number of URLs to be written to system.out. To support this
functionality, determine if a directory is safe to delete by checking if it has any
other videos apart from the one that's been downloaded.

Signed-off-by: Jason Rogena <jason@rogena.me>